refactor: move transformSolutionModel function to a new file for better organization
